requirements.txt
flask==3.0.3
werkzeug==3.0.3
bcrypt==4.1.2
PyJWT==2.9.0
email-validator==2.2.0
pytest==8.3.2
pytest-timeout==2.3.1
pytest-json-report==1.5.0
hypothesis==6.112.1
requests==2.32.3
openai==1.45.0
python-dotenv==1.0.1